pyproject.toml
[project]
name = "dyor"
version = "0.1.0"
description = "AI-powered crypto research assistant with multi-agent collaboration"
requires-python = ">=3.11,<3.13"
dependencies = [
    # LLM & Agent
    "langchain>=0.3,<0.4",
    "langgraph>=0.2,<0.4",
    "langchain-anthropic>=0.3,<0.4",
    "langchain-community>=0.3,<0.4",
    # RAG
    "chromadb>=0.5,<0.7",
    "sentence-transformers>=3.0,<4.0",
    "rank-bm25>=0.2,<0.3",
    "flashrank>=0.2,<0.3",
    "jieba>=0.42,<1.0",
    "pyyaml>=6.0,<7.0",
    # MCP
    "mcp[cli]>=1.0,<2.0",
    "httpx>=0.27,<1.0",
    # Knowledge Graph
    "networkx>=3.3,<4.0",
    # API & Frontend
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.32,<1.0",
    "websockets>=13,<15",
    "sse-starlette>=2.0,<3.0",
    "streamlit>=1.39,<2.0",
    "plotly>=5.24,<6.0",
    # Data & Utils
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    # Evaluation
    "ragas>=0.2,<0.3",
    # Logging
    "structlog>=24.0,<26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.8,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "api"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
